Make it more clear how to get help on the mappings
I found it took me a while to notice the Mapping(?) command prompt. There is not much in the readme and I could not figure out how to actually write the changes at first. This change should make it more obvious how to get help on mappings while still keeping the header concise.