- removes dependency on bash/sh -> simpler on win32 - enables automatic support for lazy.nvim's build step
The second point might be a problem, because it changes current behaviour.
`spectre_oxi` will now get built for every lazy.nvim user, even though `oxi` is not the default replace cmd.

To adress this I could:

a. Rename build.lua to build-oxi.lua.
or
b. Change the default replace cmd to `oxi` and add sample configurations for a couple of popular plugin managers to the `README.md`.